# Conflicts:
#	src/test/java/shop/itbook/itbookshop/productgroup/product/repository/elasticsearchrepository/ProductSearchRepositoryTest.java

+++ src/main/java/shop/itbook/itbookshop/book/dto/request/BookModifyRequestDto.java
@@ -41,7 +41,7 @@ public class BookModifyRequestDto {
 
     @Min(value = 0, message = "적립율은 0% 이상이어야 합니다.")
     @Max(value = 100, message = "적립율은 최대 100%입니다.")
-    private Integer increasePointPercent;
+    private Double increasePointPercent;
 
     @NotNull(message = "null을 허용하지 않습니다.")
     @Min(value = 0, message = "매입원가는 0원 이상이어야 합니다.")
